+++ R/zzz.R
@@ -1,5 +1,7 @@
 ##' @useDynLib dust, .registration = TRUE
+.onLoad <- function(...) {
+  cache$models <- simple_cache$new() # nocov
+}
 NULL
 
 cache <- new.env(parent = emptyenv())
-cache$models <- list()