+++ NEWS.md
@@ -1,6 +1,7 @@
 # dust 0.9.18
 
 * New function `dust::dust_generate` for creating a mini-package from a dust model for inspection or later loading (#204)
+* New option to `dust::dust` to skip the model cache, which may be useful when compiling with (say) different GPU options (#248)
 
 # dust 0.9.14